Refactor To Profile Card
It could be useful to have a ProfileCard used in other parts of the application.

We can create component that accepts a Person as an input property.

The 'ViewProfile' component is responsible for getting the data to pass in.

However the card now relies on data provided,
and is not responsible for fetching
the information itself.

We make use of content projection here -
this allows the consumer to control some of
what is being rendered.

Bepending on the use case - we might want
to have different actions available.