use resolver for games data
There are some times where we might want to resolve some data
before the component loads.

We can create a resolver, and configure the route to use it

When we access the data in the component, we can treat it
as if it is synchernous.

Keep in mind though, if this takes a long time - it can be harder
to display a 'loading' style indicator.

It may be better to fetch the data in the container and get more
fine grained control over when the data is diplayed / place holders / etc.